Separate mod rating render and icons into separate components.
