+++ src/components/mod_browser.tsx
@@ -11,161 +11,9 @@ import TableRow from './modbrowser/table_row';
 
 import LoadingIcon from './utils/icons/loading';
 
-import UpArrow2 from './utils/icons/up_arrow2';
-import DownArrow2 from './utils/icons/down_arrow2';
-
 type ModRowArguments = {
     mod: any
     display?: string
-};
-
-export const ModRatingRender: React.FC<ModRowArguments> = ({
-    mod
-}) => {
-    // Retrieve session.
-    const { data: session } = useSession();
-    const filters = useContext(FilterCtx);
-
-    // Retrieve rating.
-    const cur_rating = mod.ModRating[0] ?? null;
-
-    const [rating, setRating] = useState(1);
-    const [receivedRating, setReceivedRating] = useState(false);
-
-    const modRequiresUpdateMut = trpc.mod.requireUpdate.useMutation();
-
-    if (filters?.timeframe && !receivedRating) {
-        switch (filters.timeframe) {
-            case 0:
-                setRating(mod.ratingHour);
-
-                break;
-
-            case 1:
-                setRating(mod.ratingDay);
-
-                break;
-
-            case 2:
-                setRating(mod.ratingWeek);
-
-                break;
-
-            case 3:
-                setRating(mod.ratingMonth);
-
-                break;
-
-            case 4:
-                setRating(mod.ratingYear);
-
-                break;
-
-            case 5:
-                setRating(mod.totalRating);
-
-            default:
-                setRating(mod.ratingHour);
-        }
-
-        setReceivedRating(true);
-    }
-
-    // Controls whether user rated this mod or not.
-    const [didRate, setDidRate] = useState(false);
-    const [rateIsPositive, setRateIsPositive] = useState(false);
-
-    if (cur_rating && !didRate) {
-        if (cur_rating.positive)
-            setRateIsPositive(true);
-
-        setDidRate(true);
-    }
-
-    const myRatingMut = trpc.modRating.addModUserRating.useMutation();
-
-    // Arrow classes.
-    const classes_up = ["w-12", "h-12", "text-center"];
-    const classes_down = ["w-12", "h-12", "text-center"];
-
-    if (didRate) {
-        if (rateIsPositive)
-            classes_down.push("opacity-20");
-        else
-            classes_up.push("opacity-20");
-    }
-
-    return (
-        <div className="relative w-3/5 flex text-center justify-center items-center">
-            <div className="mr-1">
-                <a href="#" onClick={(e) => {
-                    e.preventDefault();
-
-                    // Submit negative rating.
-                    if (session?.user) {
-                        if (didRate && !rateIsPositive)
-                            return;
-
-                        myRatingMut.mutate({
-                            userId: session.user.id,
-                            modId: mod.id,
-                            positive: false
-                        });
-
-                        // Since we recalculate off of scheduling, set visible rating now.
-                        const curRating = Number(rating);
-                        setRating(curRating - 1);
-
-                        // Require updating.
-                        modRequiresUpdateMut.mutate({ id: mod.id });
-
-                        setDidRate(true);
-                        setRateIsPositive(false);
-                    } else if (session?.user == null)
-                        signIn("discord");
-                }}>
-                    <DownArrow2
-                        classes={classes_down}
-                    />
-                </a>
-            </div>
-            <div className="text-center">
-                <span className="text-white font-bold text-4xl">{rating.toString()}</span>
-            </div>
-            <div className="ml-1">
-                <a href="#" onClick={(e) => {
-                    e.preventDefault();
-
-                    // Submit positive rating.
-                    if (session?.user) {
-                        if (didRate && rateIsPositive)
-                            return;
-
-                        myRatingMut.mutate({
-                            userId: session.user.id,
-                            modId: mod.id,
-                            positive: true
-                        });
-
-                        // Since we recalculate off of scheduling, set visible rating now.
-                        const curRating = Number(rating);
-                        setRating(curRating + 1);
-
-                        // Require updating.
-                        modRequiresUpdateMut.mutate({ id: mod.id });
-
-                        setDidRate(true);
-                        setRateIsPositive(true);
-                    } else if (!session?.user)
-                        signIn("discord");
-                }}>
-                    <UpArrow2
-                        classes={classes_up}
-                    />
-                </a>
-            </div>
-        </div>
-    );
 };
 
 const ModRow: React.FC<ModRowArguments> = ({
