+++ src/styles/mod/rating.css
@@ -1,0 +1,15 @@
+.mod-rating-container {
+    @apply relative;
+    @apply flex;
+    @apply text-center;
+    @apply justify-center;
+    @apply items-center;
+    @apply gap-1;
+}
+
+.mod-rating-container span {
+    @apply text-white;
+    @apply font-bold;
+    @apply text-4xl;
+    @apply text-center;
+}
